# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LiveForge HTTP API
  version: 1.1.0
  description: |
    Source-aligned contract for the management API, GB28181 API, WebRTC
    WHIP/WHEP signaling, and the default internal cluster relay signaling
    paths. Management responses use `{code,message,data?}`. GB28181 handlers
    return direct JSON objects or arrays. WebRTC signaling returns SDP or
    plain-text errors.

    Management and GB28181 routes accept either an opaque bearer token from
    `api.auth` or an authenticated `lf_session` console cookie when management
    credentials are configured. If no management tokens are configured,
    compatibility mode resolves an anonymous admin principal. Health is always
    public. A missing/invalid credential returns 401; an authenticated role
    without the operation's `x-liveforge-permission` returns 403. Enabled
    per-IP rate limiting can return 429 before a handler runs.

    Internal relay signaling paths are configurable through
    `cluster.rtp.signaling_path` and `cluster.gb28181.signaling_path`; this
    document shows their defaults. Node clients read credentials atomically
    for each request, preferring `api.auth.bearer_token`, then the first named
    admin token. See `docs/recipes/cluster-relay-operations.md` before exposing
    these internal paths.
servers:
  - url: http://127.0.0.1:8090
    description: Default local management listener
security:
  - bearerAuth: []
  - consoleSession: []
...
paths:
...
  /api/v1/server/config:
    get:
      tags: [server]
      operationId: getRuntimeConfigStatus
      summary: Read redacted runtime source, version, failures, counters, and pending restart paths
      x-liveforge-permission: config:read
      responses:
        '200':
          description: Runtime configuration status
          headers:
            Cache-Control: {description: Configuration status is never cacheable, schema: {type: string, const: no-store}}
          content: {application/json: {schema: {$ref: '#/components/schemas/ConfigRuntimeStatusResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/server/config/refresh:
    post:
      tags: [server]
      operationId: refreshRuntimeConfig
      summary: Schedule an asynchronous source refresh
      description: Returns after enqueueing; source I/O remains on the background manager worker.
      x-liveforge-permission: config:reload
      responses:
        '202': {description: Refresh scheduled, content: {application/json: {schema: {$ref: '#/components/schemas/ScheduledResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/server/config/validate:
    post:
      tags: [server]
      operationId: validateRuntimeConfigDocument
      summary: Validate a complete YAML or JSON configuration document
      description: Viewer-accessible validation treats environment references literally without reading the server process environment, accepts exactly one YAML/JSON document, and rejects unknown root or nested typed fields.
      x-liveforge-permission: config:read
      requestBody:
        required: true
        content:
          application/yaml: {schema: {$ref: '#/components/schemas/ConfigYAMLDocument'}}
          application/json: {schema: {$ref: '#/components/schemas/ConfigJSONDocumentRequest'}}
      responses:
        '200': {description: Valid document, content: {application/json: {schema: {$ref: '#/components/schemas/ConfigValidationResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/server/config/apply:
    post:
      tags: [server]
      operationId: applyRuntimeConfigDocument
      summary: Persist a complete document to a writable runtime source and schedule refresh
      description: The source write completes and is serialized with source loads before the 202 response; parsing, module application, and publication remain on the background manager worker. A successful write publishes a pending desired overlay immediately so document reads remain consistent across an immediate page reload; effective values remain unchanged until asynchronous application completes. Apply and trusted runtime source loading remain permissive for source fields not mapped by the typed runtime struct. Consul KV PUT rejects redirects and never forwards X-Consul-Token. Listener, module, TLS, port, and audio-codec topology changes remain restart-required.
      x-liveforge-permission: config:reload
      parameters:
        - {$ref: '#/components/parameters/ConfigIfMatch'}
      requestBody:
        required: true
        content:
          application/yaml: {schema: {$ref: '#/components/schemas/ConfigYAMLDocument'}}
          application/json: {schema: {$ref: '#/components/schemas/ConfigJSONDocumentRequest'}}
      responses:
        '202': {description: Document written and refresh scheduled, content: {application/json: {schema: {$ref: '#/components/schemas/ConfigApplyResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Read-only source or stale document revision. No source write occurs on a revision conflict., content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/server/config/rollback:
    post:
      tags: [server]
      operationId: rollbackConfigDocument
      summary: Validate and write a retained original document, then schedule refresh
      description: Uses the original retained secrets internally. This is a conditional source write, not an immediate runtime application; restart-required and application-failure behavior matches Apply.
      x-liveforge-permission: config:reload
      parameters:
        - {name: If-Match, in: header, required: true, description: One quoted current desired document_revision., schema: {type: string}}
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [revision], properties: {revision: {type: string}}}}}
      responses:
        '202': {description: Document written and refresh scheduled, content: {application/json: {schema: {$ref: '#/components/schemas/ConfigApplyResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '428': {description: If-Match is required, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /console/logout:
    post:
      tags: [security]
      operationId: logoutConsole
      summary: Expire the browser console session cookie
      description: Sets an expired HttpOnly SameSite Strict lf_session cookie on path / and redirects to /console/login. Secure follows API TLS. This clears the browser cookie; copied signed tokens remain valid until their normal expiry or credential rotation. GET returns 405.
      security: []
      responses:
        '303': {description: Cookie expired; redirect to login}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/audit:
    get:
      tags: [security]
      operationId: listAuditEntries
      summary: Read the bounded in-memory management audit trail
      x-liveforge-permission: audit:read
      parameters:
        - {$ref: '#/components/parameters/ListLimit'}
        - {$ref: '#/components/parameters/ListOffset'}
        - {name: principal, in: query, schema: {type: string}, description: Exact principal filter.}
        - {name: action, in: query, schema: {type: string}, description: Exact action filter.}
        - {name: result, in: query, schema: {type: string}, description: Exact result filter.}
        - {name: since, in: query, schema: {type: string, format: date-time}, description: Inclusive RFC3339 lower time bound.}
        - {name: until, in: query, schema: {type: string, format: date-time}, description: Inclusive RFC3339 upper time bound.}
        - {name: order, in: query, schema: {type: string, enum: [asc, desc], default: asc}}
        - {name: format, in: query, schema: {type: string, enum: [json, ndjson], default: json}, description: NDJSON exports matching retained entries as an attachment.}
      responses:
        '200': {description: Audit entries, content: {application/json: {schema: {$ref: '#/components/schemas/AuditResponse'}}, application/x-ndjson: {schema: {type: string}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
...
  /api/v1/streams:
    get:
      tags: [streams]
      operationId: listStreams
      summary: List active streams
      x-liveforge-permission: streams:read
      parameters:
        - {name: q, in: query, required: false, description: Case-sensitive stream-key substring., schema: {type: string}}
        - {$ref: '#/components/parameters/ListLimit'}
        - {$ref: '#/components/parameters/ListOffset'}
      responses:
        '200': {description: Active streams, content: {application/json: {schema: {$ref: '#/components/schemas/StreamsResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
...
  /api/v1/sipgateway/calls:
    get:
      tags: [sipgateway]
      operationId: listSIPGatewayCalls
      summary: List calls and SIP Gateway counters
      x-liveforge-permission: sip:read
      parameters:
        - {$ref: '#/components/parameters/ListLimit'}
        - {$ref: '#/components/parameters/ListOffset'}
      responses:
        '200': {description: Calls and counters, content: {application/json: {schema: {$ref: '#/components/schemas/SIPCallsResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
    post:
      tags: [sipgateway]
      operationId: dialSIPGatewayCall
      summary: Start an outbound SIP call for an active stream
      x-liveforge-permission: sip:calls
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SIPDialRequest'}
      responses:
        '201': {description: Call created, content: {application/json: {schema: {$ref: '#/components/schemas/SIPDialResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/sipgateway/test:
    get:
      tags: [sipgateway]
      operationId: runSIPGatewaySelfTest
      summary: Run the local SIP/RTP protocol lab without a remote platform
      x-liveforge-permission: sip:read
      responses:
        '200': {description: Self-test report, content: {application/json: {schema: {$ref: '#/components/schemas/ProtocolTestResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/sipgateway/lab/sessions:
    get:
      tags: [sipgateway]
      operationId: listSIPGatewayLabSessions
      summary: List persistent local fake SIP device sessions
      x-liveforge-permission: sip:read
      responses:
        '200': {description: SIP lab sessions and playback metadata, content: {application/json: {schema: {$ref: '#/components/schemas/SIPLabSessionsResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
    post:
      tags: [sipgateway]
      operationId: startSIPGatewayLabSession
      summary: Start a persistent local fake SIP device
      description: Publish mode sends real H.264 plus PCMA or PCMU RTP/RTCP into LiveForge receivers. Receive mode accepts LiveForge's outbound INVITE, leaves the selected source stream unchanged, sends periodic per-track receiver reports, and uses the requested PCMA/PCMU target codec through the optional configured audio transcoder when the source codec differs. The configured active lab-session ceiling is enforced atomically; terminal history does not consume it and an exhausted ceiling returns 429. No external PBX is required.
      x-liveforge-permission: sip:calls
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SIPLabRequest'}}}
      responses:
        '201': {description: SIP lab session started, content: {application/json: {schema: {$ref: '#/components/schemas/SIPLabSessionResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/ProtocolLabCapacity'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/recordings:
    get:
      tags: [recording]
      operationId: listRecordings
      summary: List recordings from configured storage
      x-liveforge-permission: recordings:read
      parameters:
        - {$ref: '#/components/parameters/ListLimit'}
        - {$ref: '#/components/parameters/ListOffset'}
        - {name: q, in: query, schema: {type: string}, description: Case-sensitive recording ID or stream-key substring.}
      responses:
        '200': {description: Recordings, content: {application/json: {schema: {$ref: '#/components/schemas/RecordingsResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/gb28181/devices:
    get:
      tags: [gb28181]
      operationId: listGB28181Devices
      summary: List registered devices
      x-liveforge-permission: gb28181:read
      responses:
        '200': {description: Devices, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/GBDeviceSummary'}}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
...
  /api/v1/gb28181/channels:
    get:
      tags: [gb28181]
      operationId: listGB28181Channels
      summary: List discovered channels
      x-liveforge-permission: gb28181:read
      parameters:
        - {name: device_id, in: query, required: false, schema: {type: string}}
      responses:
        '200': {description: Channels, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/GBChannel'}}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '404': {$ref: '#/components/responses/GBNotFound'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
...
  /api/v1/gb28181/sessions:
    get:
      tags: [gb28181]
      operationId: listGB28181Sessions
      summary: List active media sessions
      x-liveforge-permission: gb28181:read
      responses:
        '200': {description: Sessions, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/GBSession'}}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
...
  /api/v1/gb28181/lab/sessions:
    get:
      tags: [gb28181]
      operationId: listGB28181LabSessions
      summary: List persistent local fake GB28181 device sessions
      x-liveforge-permission: gb28181:read
      responses:
        '200': {description: GB28181 lab sessions and playback metadata, content: {application/json: {schema: {$ref: '#/components/schemas/GBLabSessionsResponse'}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
    post:
      tags: [gb28181]
      operationId: startGB28181LabSession
      summary: Start a persistent local fake GB28181 device
      description: Publish mode performs REGISTER, Keepalive, Catalog, accepts LiveForge's server-initiated INVITE/ACK/BYE, and sends deterministic H.264 plus G.711A in PS/RTP with RTCP to LiveForge's receiver. Receive mode validates H.264 plus direct G.711A or audio that the tagged runtime can convert to G.711A and admits its subscriber before activation, then uses a module-owned outbound session with direct H.264 and an independent generation-bound target-audio reader to send PS/RTP/RTCP to the fake device; unavailable conversion fails before signaling. The configured active lab-session ceiling is enforced atomically; terminal history does not consume it and an exhausted ceiling returns 429. Subscriber admission rejection fails startup synchronously; a later outbound media failure transitions the Lab to failed with a bounded redacted last_error and releases its signaling, session, subscriber, sockets, and ports. No external GB28181 platform is required.
      x-liveforge-permission: gb28181:control
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/GBLabRequest'}}}
      responses:
        '201': {description: GB28181 lab session started, content: {application/json: {schema: {$ref: '#/components/schemas/GBLabSessionResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/ProtocolLabCapacity'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque management token}
    consoleSession: {type: apiKey, in: cookie, name: lf_session}
    streamBearerAuth: {type: http, scheme: bearer, bearerFormat: publish or subscribe token}
    streamToken: {type: apiKey, in: query, name: token}
  parameters:
    ConfigIfMatch:
      name: If-Match
      in: header
      required: false
      description: One quoted desired document_revision from GET document. A stale revision returns 409 without writing. Omission retains compatibility but concurrent changes after secret restoration are still rejected. This protects writes through this manager, not independent external source writers.
      schema: {type: string}
    ListLimit:
      name: limit
      in: query
      description: Enables bounded pagination; 1..500, defaults to 100 when only offset is supplied. Without both pagination parameters the legacy complete data shape is preserved. Filters apply first. Paginated JSON adds top-level pagination and X-Total-Count/X-Limit/X-Offset headers. Offset pages are live snapshots, not stable cursors across mutations.
      schema: {type: integer, minimum: 1, maximum: 500}
    ListOffset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
...
  responses:
...
    BadRequest: {description: Invalid or missing input, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Unauthorized: {description: Missing or invalid management credential, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Forbidden: {description: Principal lacks the required permission, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Conflict: {description: Resource state conflicts with the operation, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
...
    RateLimited: {description: Per-IP request limit exceeded, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
...
    Unavailable: {description: Required module or runtime manager is disabled, closed, or unavailable, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties: {code: {type: integer, minimum: 400}, message: {type: string}}
...
    ScheduledResponse:
      allOf: [{$ref: '#/components/schemas/ApiResponse'}, {type: object, properties: {data: {type: object, properties: {status: {type: string, const: scheduled}}}}}]
    ConfigApplyResponse:
      allOf: [{$ref: '#/components/schemas/ApiResponse'}, {type: object, properties: {data: {type: object, properties: {status: {type: string, const: written_and_refresh_scheduled}, document_revision: {type: string}}}}}]
...
    ConfigRuntimeStatusResponse: {allOf: [{$ref: '#/components/schemas/ApiResponse'}, {type: object, properties: {data: {$ref: '#/components/schemas/ConfigRuntimeStatus'}}}]}
    ConfigYAMLDocument:
      type: string
      minLength: 1
      description: Raw complete YAML document.
    ConfigJSONDocumentRequest:
      type: object
      required: [document]
      additionalProperties: false
      description: JSON envelope carrying the complete YAML/JSON document text.
      properties: {document: {type: string, minLength: 1}}
...
    ConfigValidationResponse: {allOf: [{$ref: '#/components/schemas/ApiResponse'}, {type: object, properties: {data: {$ref: '#/components/schemas/ConfigValidation'}}}]}
...
    StreamsResponse: {allOf: [{$ref: '#/components/schemas/ApiResponse'}, {type: object, properties: {data: {type: object, properties: {streams: {type: array, items: {$ref: '#/components/schemas/StreamInfo'}}}}}}]}
...
    AuditResponse: {allOf: [{$ref: '#/components/schemas/ApiResponse'}, {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/AuditEntry'}}}}]}
...
    SIPDialRequest: {type: object, additionalProperties: false, required: [target_uri, stream_key], properties: {target_uri: {type: string}, stream_key: {type: string}}}
...
    SIPCallsResponse: {allOf: [{$ref: '#/components/schemas/ApiResponse'}, {type: object, properties: {data: {type: object, properties: {calls: {type: array, items: {$ref: '#/components/schemas/SIPCall'}}, metrics: {$ref: '#/components/schemas/SIPMetrics'}}}}}]}
...